Moved Programmer project sources and headers to directories
Separated sources from headers to define a search path for the
headers inclusion. Headers from the Programmer project can be now
included with prefix that indicates their origin.
Example include <Programmer/SomeHeader.hpp>.
This will be usefull when new project containing unit tests will
be created. Rewrote all include paths to contain new prefix.
